minor smart contract data type clean up

+++ contracts/SmartBet.sol
@@ -244,7 +244,7 @@ contract SmartBet is ERC721, ChainlinkClient {
     *  @param  
     *  @return  token id
     */
-    function placeBet(uint128 _matchId, uint8 _resultBetOn)
+    function placeBet(uint256 _matchId, uint8 _resultBetOn)
         public 
         payable
         isCircuitBreakOff
@@ -306,7 +306,7 @@ contract SmartBet is ERC721, ChainlinkClient {
         return smartAssetId;
     }
     
-    function startMatch(uint128 _matchId) 
+    function startMatch(uint256 _matchId) 
         public
         onlyOwner
         matchExists(_matchId) 
@@ -324,7 +324,7 @@ contract SmartBet is ERC721, ChainlinkClient {
     *  @param  
     *  @return  success success status
     */
-    function closeMatch(uint128 _matchId, uint8 _matchResult)
+    function closeMatch(uint256 _matchId, uint8 _matchResult)
         public 
         onlyOwner
         matchExists(_matchId) 
@@ -373,7 +373,7 @@ contract SmartBet is ERC721, ChainlinkClient {
     *  @param   _smartAssetId smart asset id
     *  @return  success status
     */
-    function liquidateAsset(uint128 _smartAssetId)
+    function liquidateAsset(uint256 _smartAssetId)
         public 
         payable
         isCircuitBreakOff
